Django/project03/requirements.txt
asgiref==3.5.2
asttokens==2.0.5
backcall==0.2.0
backports.zoneinfo==0.2.1
black==22.3.0
click==8.1.3
decorator==5.1.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==4.0.5
django-filter==22.1
djangorestframework==3.13.1
executing==0.8.3
gunicorn==20.1.0
ipdb==0.13.9
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
platformdirs==2.5.2
prompt-toolkit==3.0.29
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.12.0
python-decouple==3.6
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.3.0
static3==0.7.0
toml==0.10.2
tomli==2.0.1
traitlets==5.3.0
typing-extensions==4.2.0
wcwidth==0.2.5
